saas/requirements-saas.txt
# --------------------------------------------------------------------------
# IndiaMART Lead Manager SaaS — Python Dependencies
# --------------------------------------------------------------------------

# Web framework & server
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0,<1.0

# Database (async PostgreSQL)
sqlalchemy[asyncio]>=2.0.36,<3.0
asyncpg>=0.30.0,<1.0
alembic>=1.14.0,<2.0

# Authentication & security
passlib[bcrypt]>=1.7.4,<2.0
python-jose[cryptography]>=3.3.0,<4.0
cryptography>=44.0.0,<45.0

# HTTP client (IndiaMART API calls)
httpx>=0.28.0,<1.0

# Configuration
python-dotenv>=1.0.1,<2.0
pydantic>=2.10.0,<3.0
pydantic-settings>=2.7.0,<3.0

# Task scheduling
apscheduler>=3.10.4,<4.0

# Payment gateway
razorpay>=1.4.2,<2.0

# AI / LLM
openai>=1.58.0,<2.0

# Multipart form parsing & static file serving
python-multipart>=0.0.18,<1.0
aiofiles>=24.1.0,<25.0
